requirements.txt
black
beautifulsoup4
flask
requests
pydantic
srtools
pymongo
python-dotenv
flask-cors
python-json-logger